first draft of code
