+++ apps/prairielearn/assets/stylesheets/prairielearn-ui.css
@@ -1,1 +1,2 @@
 @import url('@prairielearn/ui/components/styles.css');
+@import url('./typedPropertiesEditor.css');
